Extract controllers for primary collections
As it turns out this is just the "auth" resource.

Pulls out a controller for the auth resource so the routing will point
to it instead of the base controller. Implements its actions with the
methods "show" and "destroy", overriding the default implementation
which is only there to find this path for us at request time.
